requirements.txt
# Python packages required for the Sala Reservas application
requests
packaging
tkcalendar
mysql-connector-python
pillow
pyinstaller
